Consolidate question-card rendering into helper method
